Common/Services/ReleaseService.cs
using System;
using System.Collections.Generic;
using Etsi.Ultimate.Business;
using Etsi.Ultimate.Business.Security;
using Etsi.Ultimate.DomainClasses;
using Etsi.Ultimate.Repositories;

namespace Etsi.Ultimate.Services
{
    /// <summary>
    /// This class is the implementation in charge of all the operations concerning the releases.
    /// </summary>
    public class ReleaseService : IReleaseService
    {

        #region IReleaseService Membres

        /// <summary>
        /// Returns the list of all releases.
        /// </summary>
        /// <returns></returns>
        public KeyValuePair<List<DomainClasses.Release>,UserRightsContainer> GetAllReleases(int personID)
        {
            using (var uoW = RepositoryFactory.Resolve<IUltimateUnitOfWork>())
            {
                var releaseManager = new ReleaseManager();
                releaseManager.UoW = uoW;
                //Get list of releases
                var releases = releaseManager.GetAllReleases();
                //Get user rights
                var rightManager = ManagerFactory.Resolve<IRightsManager>();
                var personRights = rightManager.GetRights(personID);
                
                // No save needed
                return new KeyValuePair<List<Release>, UserRightsContainer>(releases, personRights);
            }
        }

        #endregion
    }
}
